+++ datafit/src/types.ts
@@ -1,38 +1,24 @@
 /**
- * Use this type to define a single-variable curve and dataset
- */
-export type SingleVariable = number;
-/**
- * Use this type to define a multi-variable curve and dataset
- */
-export type MultiVariable = Array<number>;
-/**
  * Declares whether this is a single- or multi-variable problem.
  */
-export type VariableType = SingleVariable | MultiVariable;
-/**
- * Type of function constant parameters to fit the curve with.
- */
-export type Params = Array<number>;
+export type VariableType = number | Array<number>;
 /**
  * Represents a mathematical function y = f(x) with unknown parameters.
  * @example
  * Single variable function in Typescript, 2nd degree polynomial:
  * ```ts
- * function f(x: SingleVariable, a2: number, a1: number, a0: number): number {
+ * function f(x: number, a2: number, a1: number, a0: number): number {
  *     return a2 * x ** 2 + a1 * x + a0;
  * }
  * ```
  * Multivariable function Typescript, general plane equation:
  * ```ts
- * function f(x: MultiVariable, cx: number, cy: number, c: number): number {
- *     return cx * x[0] + cy * x[1] + c;
+ * function f(x: number[], cx: number, cy: number, cz: number): number {
+ *     return cx * x[0] + cy * x[1] + cz;
  * }
  * ```
- * **Note:** `SingleVariable` can be replaced with `number` and
- * `MultiVariable` can be replaced with `Array<number>` or `number[]`.
  */
-export type fx<T extends VariableType> = (x: T, ...params: Params) => number;
+export type F<T extends VariableType> = (x: T, ...params: Array<number>) => number;
 /**
  * Stores a data point. For multivariable points, the `x`
  * coordinate contains an array of all the free variables.
@@ -48,67 +34,24 @@ export interface Datum<T extends VariableType> {
     readonly y: number;
 }
 /**
- * Contains a set of data points. Can be
- * defined to be single or multi variable
- * using the generic type qualifier. The
- * following examples are provided in
- * Typescript. Remove the type identifier
- * and qualifier for JavaScript.
- * @example
- * Single variable dataset in Typescript:
- * ```ts
- * const data: Dataset<SingleVariable> = [
- *     { x: 1, y: 1 },
- *     { x: 2, y: 3 },
- *     { x: 3, y: 4 },
- * ];
- * ```
- * Multivariable dataset in Typescript:
- * ```ts
- * const data: Dataset<MultiVariable> = [
- *     { x: [0, 0], y: 1 },
- *     { x: [1, 0], y: 3 },
- *     { x: [0, 1], y: 4 },
- *     { x: [1, 1], y: 5 },
- * ];
- * ```
- */
-export type Dataset<T extends VariableType> = Array<Datum<T>>;
-/**
  * Includes information about a best-fit for a curve.
  */
-export interface Fit {
+export interface Summary {
     /**
      * Contains the set of best-fit parameters for the function `f(x)`
      */
-    readonly params: Params;
+    readonly params: Array<number>;
     /**
      * This is the residual sum of squared errors.
      */
-    readonly err: number;
-}
-/**
- * Configuration options for curve fitting.
- */
-export interface Config {
+    readonly error: number;
     /**
-     * Determines the number of generations, or iterations.
+     * The number of data points given to the model.
      */
-    readonly generations: number;
+    readonly Ndata: number;
     /**
-     * Determines the number of parameters sets to generate.
+     * The average absolute error per data point, comparing the given
+     * dataset to the model output with the set of best-fit parameters.
      */
-    readonly population: number;
-    /**
-     * Determines how many survivors remain after every generation.
-     */
-    readonly survivors: number;
-    /**
-     * Determines how much a set of parameters can mutate on the first generation.
-     */
-    readonly initialDeviation: number;
-    /**
-     * Determines how much a set of parameters can mutate on the final generation.
-     */
-    readonly finalDeviation: number;
+    readonly avgAbsErr: number;
 }